fixed minor warning

+++ src/bstone_in_manager.mm
@@ -83,7 +83,7 @@ void InputManager::AddRectRegionEvent(float x, float y, float width, float heigh
 
 void InputManager::ClearRegionEvents(void)
 {
-    int size = m_RegionEventVector.size();
+    int size = static_cast<int>(m_RegionEventVector.size());
 
     for (int i = 0; i < size; i++)
     {
